修复已知bug；没有订货会价格，使用客户默认价格体系对应价格

+++ app/src/main/java/com/yifarj/yifadinghuobao/model/entity/MettingLoginEntity.java
@@ -97,9 +97,9 @@ public class MettingLoginEntity {
         public int vip_sex;
         public int card_status;
         public int oper_id;
-        public int acc_num;
-        public int dec_num;
-        public int res_num;
+        public double acc_num;
+        public double dec_num;
+        public double res_num;
         public String memo;
         public String homeplace;
         public String marriage;
